requirements.txt
alembic==1.13.2
fastapi[standard]==0.114.1
pwdlib[argon2]==0.2.1
pydantic-settings==2.5.2
pyjwt==2.9.0